Config Review Complete
Config files were brought into the project, to be managed by GitHub.
All the config files will need to be rewritten according to the new
standard. (for a later commit)

All calls to the config have now migrated to the OBConfigManager
MainActivity has been cleaned of config related functions and static
Strings

OBB Expansion Manager has been removed, as it’s no longer a viable
solution for the project.
It has been replaced by the external assets path and search function in
the OBSystemsManager.

From a path in the config file, it will look for any removable drive,
mounted partition, /sdcard and /data folders.